// frontend/src/services/api/constants.ts
export enum MessageTypes {
  USER_REGISTER = 'USER_REGISTER',
  USER_LOGIN = 'USER_LOGIN',
  GET_USERS = 'GET_USERS',
  GET_USER = 'GET_USER',
  GET_CURRENT_USER = 'GET_CURRENT_USER',
  CREATE_BOARD = 'CREATE_BOARD',
  GET_BOARD = 'GET_BOARD',
  GET_BOARDS = 'GET_BOARDS',
  GET_BOARD_USERS_QUANTITY = 'GET_BOARD_USERS_QUANTITY',
  USER_ENTER_BOARD = 'USER_ENTER_BOARD',
  USER_LEAVE_BOARD = 'USER_LEAVE_BOARD',
  POINT_DRAW = 'POINT_DRAW',
}